Establish separate fork (#11)
Migrated to package social.orbitearth.opengrape. Edited POM and reset
version 1.0.0. Updated README.
