+++ pom.xml
@@ -4,14 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>io.github.kientux</groupId>
+    <groupId>social.orbitearth</groupId>
     <artifactId>opengrape</artifactId>
-    <version>1.0.4</version>
+    <version>1.0.0</version>
     <packaging>jar</packaging>
 
     <name>${project.groupId}:${project.artifactId}</name>
-    <description>Java wrapper of Open Graph protocol</description>
-    <url>https://github.com/kientux/opengrape</url>
+    <description>An Open Graph parser, in Java</description>
+    <url>https://github.com/charleshope/opengrape</url>
 
     <licenses>
         <license>
@@ -22,17 +22,16 @@
 
     <developers>
         <developer>
+            <name>Charles Hope</name>
+            <email>charles@orbitearth.social</email>
+            <url>https://github.com/charleshope</url>
+        </developer>
+        <developer>
             <name>Kien Nguyen</name>
             <email>ntkien93@gmail.com</email>
             <url>https://github.com/kientux</url>
         </developer>
     </developers>
-
-    <scm>
-        <connection>scm:git:git://github.com/kientux/opengrape.git</connection>
-        <developerConnection>scm:git:ssh://github.com:kientux/opengrape.git</developerConnection>
-        <url>https://github.com/kientux/opengrape/tree/main</url>
-    </scm>
 
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
@@ -65,17 +64,6 @@
         </dependency>
     </dependencies>
 
-    <distributionManagement>
-        <snapshotRepository>
-            <id>ossrh</id>
-            <url>https://s01.oss.sonatype.org/content/repositories/snapshots</url>
-        </snapshotRepository>
-        <repository>
-            <id>ossrh</id>
-            <url>https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/</url>
-        </repository>
-    </distributionManagement>
-
     <build>
         <plugins>
             <plugin>
@@ -93,46 +81,8 @@
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-javadoc-plugin</artifactId>
-                <version>3.3.2</version>
-                <executions>
-                    <execution>
-                        <id>attach-javadocs</id>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-gpg-plugin</artifactId>
-                <version>3.0.1</version>
-                <executions>
-                    <execution>
-                        <id>sign-artifacts</id>
-                        <phase>verify</phase>
-                        <goals>
-                            <goal>sign</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>2.22.0</version>
-            </plugin>
-            <plugin>
-                <groupId>org.sonatype.plugins</groupId>
-                <artifactId>nexus-staging-maven-plugin</artifactId>
-                <version>1.6.7</version>
-                <extensions>true</extensions>
-                <configuration>
-                    <serverId>ossrh</serverId>
-                    <nexusUrl>https://s01.oss.sonatype.org/</nexusUrl>
-                    <autoReleaseAfterClose>true</autoReleaseAfterClose>
-                </configuration>
             </plugin>
         </plugins>
     </build>
