+++ src/main/java/social/orbitearth/opengrape/OpenGrapeResponseException.java
@@ -1,4 +1,4 @@
-package io.github.kientux;
+package social.orbitearth.opengrape;
 
 public class OpenGrapeResponseException extends Exception {
     public static class UnexpectedStatusCode extends OpenGrapeResponseException {
